Moves date and option picker from catalog item result screen to next screen.
